Shift CDAuctioneer to use new CDFacility interface

+++ src/test/mocks/MockConvertibleDepositAuctioneer.sol
@@ -40,7 +40,7 @@ contract MockConvertibleDepositAuctioneer is IConvertibleDepositAuctioneer, Poli
         uint256 newMinPrice
     ) external override returns (uint256 remainder) {}
 
-    function setTimeToExpiry(uint256 newTime) external override {}
+    function setTimeToExpiry(uint48 newTime) external override {}
 
     function setTickStep(uint256 newStep) external override {}
 }
